// src/types/filter.types.ts
export interface FiltersType {
  age?: string;
  height?: string;
  weight?: string;
  monthlyIncome?: string;
  lookingFor?: string;
  religion?: string;
  country?: string;
  education?: string;
  profession?: string;
  familyMember?: string;
  languageSpoken?: string;
  politicalView?: string;
  maritalStatus?: string;
  hasChildren?: string;
  hasPet?: boolean;
  dietaryPreference?: string;
  smokingHabit?: string;
  drinkingHabit?: string;
}